refactor(sec): extract shared EDGAR XML submission parsing from filing processors (#2875)
