+++ src/Equibles.Sec.HostedService/Services/Form144FilingProcessor.cs
@@ -1,15 +1,14 @@
 using System.Globalization;
-using System.Text.RegularExpressions;
 using System.Xml.Linq;
 using Equibles.CommonStocks.Data.Models;
 using Equibles.Errors.BusinessLogic;
-using Equibles.Errors.Data.Models;
 using Equibles.InsiderTrading.Data.Models;
 using Equibles.InsiderTrading.Repositories;
 using Equibles.Integrations.Sec.Contracts;
 using Equibles.Integrations.Sec.Models;
 using Equibles.Sec.Data.Models;
 using Equibles.Sec.HostedService.Contracts;
+using Equibles.Sec.HostedService.Helpers;
 using Microsoft.EntityFrameworkCore;
 
 namespace Equibles.Sec.HostedService.Services;
@@ -70,7 +69,15 @@ public class Form144FilingProcessor : IFilingProcessor
             return false;
         }
 
-        var root = await TryParseSubmission(content, filing, companyTicker);
+        var root = await EdgarXmlSubmissionParser.TryParseSubmission(
+            content,
+            filing,
+            companyTicker,
+            "Form 144",
+            "Form144.ParseXml",
+            _logger,
+            _errorReporter
+        );
         if (root == null)
             return false;
 
@@ -97,49 +104,6 @@ public class Form144FilingProcessor : IFilingProcessor
         );
 
         return true;
-    }
-
-    private async Task<XElement> TryParseSubmission(
-        string content,
-        FilingData filing,
-        string companyTicker
-    )
-    {
-        var sanitized = SanitizeXml(content);
-
-        // Form 144 has only been filed as XML since the SEC's April 2022 e-filing mandate,
-        // so a submission without the <edgarSubmission> root is unexpected — skip quietly.
-        if (!sanitized.Contains("<edgarSubmission", StringComparison.OrdinalIgnoreCase))
-        {
-            _logger.LogDebug(
-                "Skipping non-XML Form 144 filing for {Ticker} - {AccessionNumber}",
-                companyTicker,
-                filing.AccessionNumber
-            );
-            return null;
-        }
-
-        try
-        {
-            return XDocument.Parse(sanitized).Root;
-        }
-        catch (System.Xml.XmlException ex)
-        {
-            _logger.LogWarning(
-                ex,
-                "Malformed Form 144 XML for {Ticker} - {AccessionNumber}",
-                companyTicker,
-                filing.AccessionNumber
-            );
-            await _errorReporter.Report(
-                ErrorSource.DocumentScraper,
-                "Form144.ParseXml",
-                ex.Message,
-                ex.StackTrace,
-                $"ticker: {companyTicker}, accession: {filing.AccessionNumber}"
-            );
-            return null;
-        }
     }
 
     private static Form144Filing ParseFiling(XElement root, Guid companyId, FilingData filing)
@@ -221,23 +185,8 @@ public class Form144FilingProcessor : IFilingProcessor
         return string.IsNullOrEmpty(value) ? null : value;
     }
 
-    private const string XmlEnvelopeStart = "<XML>";
-    private const string XmlEnvelopeEnd = "</XML>";
-
-    internal static string SanitizeXml(string xml)
-    {
-        // SEC filings wrap the actual XML inside an SGML envelope (one .txt holds the whole
-        // submission); pull out the <XML>...</XML> body before parsing.
-        var xmlStart = xml.IndexOf(XmlEnvelopeStart, StringComparison.OrdinalIgnoreCase);
-        var xmlEnd = xml.IndexOf(XmlEnvelopeEnd, StringComparison.OrdinalIgnoreCase);
-        if (xmlStart >= 0 && xmlEnd > xmlStart)
-        {
-            xml = xml[(xmlStart + XmlEnvelopeStart.Length)..xmlEnd].Trim();
-        }
-
-        // Escape stray ampersands that aren't already part of an entity.
-        return Regex.Replace(xml, @"&(?!(amp|lt|gt|quot|apos|#\d+|#x[\da-fA-F]+);)", "&amp;");
-    }
+    // Pinned by processor-scoped tests; the implementation lives in the shared parser.
+    internal static string SanitizeXml(string xml) => EdgarXmlSubmissionParser.SanitizeXml(xml);
 
     // Form 144 dates are US-format MM/dd/yyyy. Parse culture-independently so a non-Gregorian
     // host culture doesn't silently drop every date.
